moves chart types into types
