--- src/spotify/chart/config.ts
@@ -1,1 +1,0 @@
-export const config = {}
