+++ src/spotify/index.ts
@@ -2,4 +2,5 @@ export * from './listens'
 export * from './plays'
 export * from './types'
 export * from './utils'
+export * from './chart'
 export { SpotifyHistory } from './spotify_data'
