+++ src/spotify/types/index.ts
@@ -1,2 +1,3 @@
 export * from './Song'
 export * from './History'
+export * from './Chart'
